Change error handling when starting a timeout task, switch command counter from integer to long, and optimize removing timeout commands

+++ src/main/java/com/microsoft/sqlserver/jdbc/TimeoutCommand.java
@@ -6,7 +6,7 @@
 package com.microsoft.sqlserver.jdbc;
 
 import java.util.concurrent.ScheduledFuture;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 
 /**
@@ -18,7 +18,7 @@ abstract class TimeoutCommand<T> {
     private final T command;
     private final SQLServerConnection sqlServerConnection;
     private ScheduledFuture<?> timeoutTask;
-    static AtomicInteger uniqueId = new AtomicInteger();
+    static AtomicLong uniqueId = new AtomicLong();
     private final long id;
 
     TimeoutCommand(int timeout, T command, SQLServerConnection sqlServerConnection) {
